refactor(entity): 修改目录文件名
(1)修改目录文件名
