+++ addons/@/apps/jm-banner/src/main/java/cn/jimoos/constant/BannerStatus.java
@@ -1,4 +1,4 @@
-package cn.jimoos.dic;
+package cn.jimoos.constant;
 
 /**
  * @author :keepcleargas
