+++ addons/@/apps/jm-order/src/main/java/cn/jimoos/constant/OrderType.java
@@ -1,4 +1,4 @@
-package cn.jimoos.dic;
+package cn.jimoos.constant;
 
 /**
  * @author :keepcleargas
