+++ starter/cron-starter/src/main/java/cn/jimoos/cron/OrderJob.java
@@ -2,7 +2,7 @@ package cn.jimoos.cron;
 
 import cn.jimoos.config.OrderProperties;
 import cn.jimoos.dao.OrderMapper;
-import cn.jimoos.dic.OrderStatus;
+import cn.jimoos.constant.OrderStatus;
 import cn.jimoos.form.order.CancelForm;
 import cn.jimoos.model.Order;
 import cn.jimoos.service.OrderComposeService;
